Use CompletableFutures.allAsList instead of manually computing the partitions list.
