+++ kafka-rest/src/main/java/io/confluent/kafkarest/controllers/ClusterManagerImpl.java
@@ -20,6 +20,7 @@ import static java.util.Collections.singletonList;
 import static java.util.Collections.unmodifiableList;
 import static java.util.Objects.requireNonNull;
 
+import io.confluent.kafkarest.common.KafkaFutures;
 import io.confluent.kafkarest.entities.Broker;
 import io.confluent.kafkarest.entities.Cluster;
 import io.confluent.kafkarest.exceptions.UnsupportedProtocolException;
