+++ kafka-rest/src/main/java/io/confluent/kafkarest/controllers/ReplicaManagerImpl.java
@@ -20,6 +20,8 @@ import static io.confluent.kafkarest.controllers.Entities.findEntityByKey;
 import static java.util.Collections.singletonList;
 import static java.util.Objects.requireNonNull;
 
+import io.confluent.kafkarest.common.CompletableFutures;
+import io.confluent.kafkarest.common.KafkaFutures;
 import io.confluent.kafkarest.entities.Partition;
 import io.confluent.kafkarest.entities.PartitionReplica;
 import java.util.List;
