--- kafka-rest/src/test/java/io/confluent/kafkarest/CompletableFutures.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 Confluent Inc.
- *
- * Licensed under the Confluent Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
- *
- * http://www.confluent.io/confluent-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package io.confluent.kafkarest;
-
-import java.util.concurrent.CompletableFuture;
-
-/**
- * Utilities to deal with {@link CompletableFuture}.
- */
-public final class CompletableFutures {
-
-  private CompletableFutures() {
-  }
-
-  /**
-   * Returns a {@link CompletableFuture} that is completed exceptionally with the given {@code
-   * exception}.
-   */
-  public static <T> CompletableFuture<T> failedFuture(Throwable exception) {
-    CompletableFuture<T> future = new CompletableFuture<>();
-    future.completeExceptionally(exception);
-    return future;
-  }
-}
