+++ kafka-rest/src/test/java/io/confluent/kafkarest/TestUtils.java
@@ -24,8 +24,6 @@ import io.confluent.kafkarest.entities.ProduceRecord;
 import io.confluent.kafkarest.entities.v2.PartitionOffset;
 import io.confluent.rest.entities.ErrorMessage;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
@@ -41,7 +39,6 @@ import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.common.KafkaFuture;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -289,10 +286,5 @@ public class TestUtils {
     consumerConfig.setProperty(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, valueDeserializerClassName);
     consumerConfig.putAll(deserializerProps);
     return new KafkaConsumer<>(consumerConfig);
-  }
-
-  public static <T> KafkaFuture<T> failedFuture(RuntimeException exception) {
-    KafkaFuture<T> future = KafkaFuture.completedFuture(null);
-    return future.whenComplete((value, error) -> { throw exception; });
   }
 }
