+++ kafka-rest/src/test/java/io/confluent/kafkarest/controllers/ClusterManagerImplTest.java
@@ -15,7 +15,7 @@
 
 package io.confluent.kafkarest.controllers;
 
-import static io.confluent.kafkarest.TestUtils.failedFuture;
+import static io.confluent.kafkarest.common.KafkaFutures.failedFuture;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static org.easymock.EasyMock.anyObject;
