+++ kafka-rest/src/test/java/io/confluent/kafkarest/resources/v3/BrokerConfigResourceTest.java
@@ -1,6 +1,6 @@
 package io.confluent.kafkarest.resources.v3;
 
-import static io.confluent.kafkarest.CompletableFutures.failedFuture;
+import static io.confluent.kafkarest.common.CompletableFutures.failedFuture;
 import static java.util.concurrent.CompletableFuture.completedFuture;
 import static org.easymock.EasyMock.expect;
 import static org.easymock.EasyMock.replay;
